move minmax tree into submodule for better encapsulation
